+++ src/lib/index.ts
@@ -1,3 +1,3 @@
-export { LibComponent } from './src/component/lib.component';
-export { LibService } from './src/service/lib.service';
-export { LibModule } from './src/module';
+export { AngularFontAwesomeComponent } from './src/component/angular-font-awesome.component';
+export { AngularFontAwesomeService } from './src/service/angular-font-awesome.service';
+export { AngularFontAwesomeModule } from './src/angular-font-awesome.module';