+++ src/lib/src/component/angular-font-awesome.component.spec.ts
@@ -2,22 +2,22 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 
-import { LibComponent } from './lib.component';
+import { AngularFontAwesomeComponent } from './angular-font-awesome.component';
 
-describe('LibComponent', function () {
+describe('AngularFontAwesomeComponent', function () {
   let de: DebugElement;
-  let comp: LibComponent;
-  let fixture: ComponentFixture<LibComponent>;
+  let comp: AngularFontAwesomeComponent;
+  let fixture: ComponentFixture<AngularFontAwesomeComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [LibComponent]
+      declarations: [AngularFontAwesomeComponent]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(LibComponent);
+    fixture = TestBed.createComponent(AngularFontAwesomeComponent);
     comp = fixture.componentInstance;
     de = fixture.debugElement.query(By.css('h2'));
   });