+++ src/lib/src/service/angular-font-awesome.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { AngularFontAwesomeService } from './angular-font-awesome.service';
+
+describe('AngularFontAwesomeService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [AngularFontAwesomeService]
+    });
+  });
+
+  it('should create service', inject([AngularFontAwesomeService], (service: AngularFontAwesomeService) => {
+    expect(service).toBeTruthy();
+  }));
+});